backend/requirements.txt
# Ocean Sentinel API - Dependencies
# Optimisé pour faible consommation RAM

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
